requirements.txt
psutil>=5.8.0
pyautogui>=0.9.53
keyboard>=0.13.5
Pillow>=9.1.0
torch>=1.7.0
torchvision>=0.8.1
opencv-python>=4.5.5.64
pywin32>=302
# YOLOv5 requirements
# Usage: pip install -r requirements.txt

# YOLOv5 Base ------------------------------------------------------------------------
gitpython>=3.1.30
matplotlib>=3.3
numpy>=1.18.5
opencv-python>=4.1.1
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
thop>=0.1.1  # FLOPs computation
tqdm>=4.64.0
ultralytics>=8.0.100

# Plotting --------------------------------------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

# Deploy ----------------------------------------------------------------------
setuptools>=65.5.1 # Snyk vulnerability fix